==> 02e_opioidsmanuf_classification_logit.R <==
## "The Economic Underpinnings of the Drug Epidemic" - Nathan Seltzer
## DOI: https://doi.org/10.1016/j.ssmph.2020.100679

## Title: Corrected Opioid Logit Predictions 
## File: 02e_opioidsmanuf_classification_logit.R
## Description: Creates logit program to predict unspecified opioid-related drug deaths 


## READ IN DATA
mort_full3 <- fread( "V:/seltzer/mortality/data/full_mortality_data_1999_2017_mort_full3_-RESTRICTED_keeponwinstat.csv")

`%notin%` <- Negate(`%in%`)

## Make edits
mort_full3_unk <- mort_full3 %>%
  filter(ucod == "X40" |
           ucod == "X41" |
           ucod == "X42" |
           ucod == "X43" |
           ucod == "X44" |
           ucod == "X60" |
           ucod == "X61" |
           ucod == "X62" |
           ucod == "X63" |
           ucod == "X64" |
           ucod == "X85" |
           ucod == "Y10" |
           ucod == "Y11" |
           ucod == "Y12" |
           ucod == "Y13" | 
           ucod == "Y14" ) %>%
  mutate(opioid.death = opioid) %>%
  mutate(T509 = ifelse( record_1	==	"T509" |			
                           record_2	==	"T509" |			
                           record_3	==	"T509" |			
                           record_4	==	"T509" |			
                           record_5	==	"T509" |			
                           record_6	==	"T509" |			
                           record_7	==	"T509" |			
                           record_8	==	"T509" |			
                           record_9	==	"T509" |			
                           record_10	==	"T509" |			
                           record_11	==	"T509" |			
                           record_12	==	"T509" |			
                           record_13	==	"T509" |			
                           record_14	==	"T509" |			
                           record_15	==	"T509" |			
                           record_16	==	"T509" |			
                           record_17	==	"T509" |			
                           record_18	==	"T509" |			
                           record_19	==	"T509" |			
                           record_20	==	"T509" ,	1,	0	),
         hisp = ifelse(hispanic == 0 | hispanic == 100, "non-Hispanic", "Hispanic")) %>%
  mutate(nonop = ifelse( record_1	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |
                          record_2	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_3	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_4	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_5	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_6	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_7	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_8	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_9	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_10	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_11	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_12	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_13	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_14	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_15	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_16	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_17	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_18	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_19	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") |			
                          record_20	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") 
                         ,	1,	0	),
         nonoponly = ifelse((record_1	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_1 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |
                              (record_2	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_2 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_3	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_3 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_4	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_4 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_5	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_5 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_6	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_6 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_7	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_7 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_8	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_8 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_9	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_9 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_10	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_10 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_11	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_11 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_12	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_12 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_13	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_13 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_14	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_14 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_15	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_15 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_16	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_16 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_17	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_17 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_18	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_18 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_19	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_19 %notin% c("T400", "T401", "T402", "T403", "T404", "T406")) |			
                              (record_20	%in% c("T360" , "T361" , "T362" , "T363" , "T364" , "T365" , "T366" , "T367" , "T368" , "T369" , "T370" , "T371" , "T372" , "T373" , "T374" , "T375" , "T378" , "T379" , "T380" , "T381" , "T382" , "T383" , "T384" , "T385" , "T386" , "T387" , "T388" , "T389" , "T390" , "T391" , "T392" , "T393" , "T394" , "T398" , "T399" , "T405" , "T407" , "T408" , "T409" , "T410" , "T411" , "T412" , "T413" , "T414" , "T415" , "T420" , "T421" , "T422" , "T423" , "T424" , "T425" , "T426" , "T427" , "T428" , "T430" , "T431" , "T432" , "T433" , "T434" , "T435" , "T436" , "T438" , "T439" , "T440" , "T441" , "T442" , "T443" , "T444" , "T445" , "T446" , "T447" , "T448" , "T449" , "T450" , "T451" , "T452" , "T453" , "T454" , "T455" , "T456" , "T457" , "T458" , "T459" , "T460" , "T461" , "T462" , "T463" , "T464" , "T465" , "T466" , "T467" , "T468" , "T469" , "T470" , "T471" , "T472" , "T473" , "T474" , "T475" , "T476" , "T477" , "T478" , "T479" , "T480" , "T481" , "T482" , "T483" , "T484" , "T485" , "T486" , "T487" , "T490" , "T491" , "T492" , "T493" , "T494" , "T495" , "T496" , "T497" , "T498" , "T499" , "T500" , "T501" , "T502" , "T503" , "T504" , "T505" , "T506" , "T507" , "T508") & record_20 %notin% c("T400", "T401", "T402", "T403", "T404", "T406"))  
                             ,	1,	0	))
  
mort_full3_unk$T509[is.na(mort_full3_unk$T509)] <- 0
mort_full3_unk$opioid.death[is.na(mort_full3_unk$opioid.death)] <- 0
  
head(mort_full3_unk)
table(mort_full3_unk$T509, mort_full3_unk$opioid.death)




nonop <- read.csv("V:/seltzer/mortality/prg/leading causes.csv")




unk <-  mort_full3_unk %>%
  mutate(opioid.classification = ifelse(opioid.death == 1, 1, 0),
         nonopioid.classification = ifelse(opioid.death == 0 & nonop == 1, 1, 0),
         unclassified.classification = ifelse(opioid.death == 0 & nonop == 0, 1, 0),
         placdth = ifelse(placdth == 9, NA, placdth),
         agecat = ifelse(agecat == "", NA, agecat),
         race_r = ifelse(race == 1, "white",
                         ifelse(race == 2, "black",
                                ifelse(race == 3, "ai",
                                       ifelse(race %in% c(3:78), "api", NA)))),
         weekday = ifelse(weekday == 9, NA, weekday))



##


opioidmodel <- function(df, yr) {
  
  data <- filter(df, year == yr & unclassified.classification != 1)
  
  # data <- unk2016
  # data <- NULL
  
  smp_size <- floor(0.80 * nrow(data))
  
  ## set the seed to make your partition reproducible
  set.seed(151)
  train_ind <- sample(seq_len(nrow(data)), size = smp_size)
  
  train <- data[train_ind, ]
  test <- data[-train_ind, ]
  
  
  
  train$opioid.death <- factor(train$opioid.death)
  logitmodel <- glm(opioid.death ~ factor(agecat) + factor(race_r) + factor(sex) + factor(hisp) + factor(marstat) + factor(weekday) + factor(monthdth) + factor(placdth) + factor(state_fips ), 
            data = train, 
            family=binomial(link="logit"))
  summary(logitmodel)
  
  
  train$prob.train <- logitmodel %>%
    predict(train, type = "response")
  
  train$pred.train <- ifelse(train$prob.train > 0.5, 1, 0)
  
  test$prob.test <- logitmodel %>%
    predict(test, type = "response")
  
  test$pred.test <- ifelse(test$prob.test > 0.5, 1, 0)
  
  
  ##get accuracy
  
  cm_train <- table(train$opioid.death, train$pred.train)
  
  acc_train <- (cm_train[1,1] + cm_train[2,2] )/ (cm_train[1,1] + cm_train[2,2] + cm_train[1,2] + cm_train[2,1] )
  
 
  
  cm_test <- table(test$opioid.death, test$pred.test)

  acc_test <- (cm_test[1,1] + cm_test[2,2] )/ (cm_test[1,1] + cm_test[2,2] + cm_test[1,2] + cm_test[2,1] )
  
  ## predict deaths of unclassified
  
  corrected_data <- filter(df, year == yr)
  
  corrected_data$prob <- logitmodel %>%
    predict(corrected_data, type = "response")
  
  corrected_data$pred <- ifelse(corrected_data$prob > 0.5, 1, 0)
  
  corrected_data$prob_mix = ifelse(corrected_data$unclassified.classification == 1, corrected_data$prob, corrected_data$opioid.death)
  corrected_data$pred_mix = ifelse(corrected_data$unclassified.classification == 1, corrected_data$pred, corrected_data$opioid.death)

  
  ##total deaths
  
  total <- summarize(corrected_data, 
                     alldrug = sum(death),
                      opioid = sum(opioid.death),
                      opioid_plus_predicted = sum(pred_mix, na.rm = TRUE))
  
  
  elements <- list(train, test, logitmodel, cm_train, acc_train, cm_test, acc_test, corrected_data, total)
  
  # 1. train dataset
  # 2. test dataset
  # 3. logit model
  # 4. confusion matrix train
  # 5. predicted accuracy train
  # 6. confusion matrix test
  # 7. predicted accuracy train
  # 8. corrected dataset with predictions
  # 9. totals
  
  return(elements)
  
}
  

pred_1999	<- opioidmodel(unk, 	1999	)
pred_2000	<- opioidmodel(unk, 	2000	)
pred_2001	<- opioidmodel(unk, 	2001	)
pred_2002	<- opioidmodel(unk, 	2002	)
pred_2003	<- opioidmodel(unk, 	2003	)
pred_2004	<- opioidmodel(unk, 	2004	)
pred_2005	<- opioidmodel(unk, 	2005	)
pred_2006	<- opioidmodel(unk, 	2006	)
pred_2007	<- opioidmodel(unk, 	2007	)
pred_2008	<- opioidmodel(unk, 	2008	)
pred_2009	<- opioidmodel(unk, 	2009	)
pred_2010	<- opioidmodel(unk, 	2010	)
pred_2011	<- opioidmodel(unk, 	2011	)
pred_2012	<- opioidmodel(unk, 	2012	)
pred_2013	<- opioidmodel(unk, 	2013	)
pred_2014	<- opioidmodel(unk, 	2014	)
pred_2015	<- opioidmodel(unk, 	2015	)
pred_2016	<- opioidmodel(unk, 	2016	)
pred_2017	<- opioidmodel(unk, 	2017	)





a <- opioidmodel(unk, 2010)


a[[5]]
a[[7]]
table(a[[8]]$pred_mix)
table(a[[8]]$opioid.death)


pred_1999[[7]]
pred_2007[[7]]
pred_2017[[7]]



b <- data.frame(
  year = 1999:2017,
  test_acc =c (pred_1999[[7]],
               pred_2000[[7]],
               pred_2001[[7]],
               pred_2002[[7]],
               pred_2003[[7]],
               pred_2004[[7]],
               pred_2005[[7]],
               pred_2006[[7]],
               pred_2007[[7]],
               pred_2008[[7]],
               pred_2009[[7]],
               pred_2010[[7]],
               pred_2011[[7]],
               pred_2012[[7]],
               pred_2013[[7]],
               pred_2014[[7]],
               pred_2015[[7]],
               pred_2016[[7]],
               pred_2017[[7]]),
  train_acc =c (pred_1999[[5]],
               pred_2000[[5]],
               pred_2001[[5]],
               pred_2002[[5]],
               pred_2003[[5]],
               pred_2004[[5]],
               pred_2005[[5]],
               pred_2006[[5]],
               pred_2007[[5]],
               pred_2008[[5]],
               pred_2009[[5]],
               pred_2010[[5]],
               pred_2011[[5]],
               pred_2012[[5]],
               pred_2013[[5]],
               pred_2014[[5]],
               pred_2015[[5]],
               pred_2016[[5]],
               pred_2017[[5]]),
  alldrug   =c (pred_1999[[9]][[1]],
                pred_2000[[9]][[1]],
                pred_2001[[9]][[1]],
                pred_2002[[9]][[1]],
                pred_2003[[9]][[1]],
                pred_2004[[9]][[1]],
                pred_2005[[9]][[1]],
                pred_2006[[9]][[1]],
                pred_2007[[9]][[1]],
                pred_2008[[9]][[1]],
                pred_2009[[9]][[1]],
                pred_2010[[9]][[1]],
                pred_2011[[9]][[1]],
                pred_2012[[9]][[1]],
                pred_2013[[9]][[1]],
                pred_2014[[9]][[1]],
                pred_2015[[9]][[1]],
                pred_2016[[9]][[1]],
                pred_2017[[9]][[1]]),
  opioid   =c (pred_1999[[9]][[2]],
                pred_2000[[9]][[2]],
                pred_2001[[9]][[2]],
                pred_2002[[9]][[2]],
                pred_2003[[9]][[2]],
                pred_2004[[9]][[2]],
                pred_2005[[9]][[2]],
                pred_2006[[9]][[2]],
                pred_2007[[9]][[2]],
                pred_2008[[9]][[2]],
                pred_2009[[9]][[2]],
                pred_2010[[9]][[2]],
                pred_2011[[9]][[2]],
                pred_2012[[9]][[2]],
                pred_2013[[9]][[2]],
                pred_2014[[9]][[2]],
                pred_2015[[9]][[2]],
                pred_2016[[9]][[2]],
                pred_2017[[9]][[2]]),  
  opioid_plus_predicted   =c (pred_1999[[9]][[3]],
                pred_2000[[9]][[3]],
                pred_2001[[9]][[3]],
                pred_2002[[9]][[3]],
                pred_2003[[9]][[3]],
                pred_2004[[9]][[3]],
                pred_2005[[9]][[3]],
                pred_2006[[9]][[3]],
                pred_2007[[9]][[3]],
                pred_2008[[9]][[3]],
                pred_2009[[9]][[3]],
                pred_2010[[9]][[3]],
                pred_2011[[9]][[3]],
                pred_2012[[9]][[3]],
                pred_2013[[9]][[3]],
                pred_2014[[9]][[3]],
                pred_2015[[9]][[3]],
                pred_2016[[9]][[3]],
                pred_2017[[9]][[3]]))

  
  
library(viridis)
library(scico)
library(ggthemes)


ggplot(b, aes(x = year, y = test_acc)) + 
  # geom_point(size = 1.25) + 
  geom_smooth(se = FALSE, size = 1.25, color = "dodgerblue3") +
  geom_point(size = 1.25, color = "black") +
  scale_y_continuous(breaks=seq(.65,.8,.5)) +
  ylim(.61,.809) +
  theme_tufte() +
  xlab("Year") +
  ylab("Predictive Accuracy Rate") 
ggsave(device = "png", 
       filename = "V:/seltzer/mortality/output/PredictiveAccuracy.png",
       width = 5,
       height = 5,
       dpi = 700)


b2 <- gather(b, "measure", "value", 2:6) %>%
  filter(measure %in% c("alldrug", "opioid", "opioid_plus_predicted"))



b2$measure <- factor(b2$measure , levels = c("alldrug","opioid_plus_predicted","opioid"),
                     labels = c("Total Drug Overdoses", 
                                "Specified Opioid Overdoses + Predicted Opioid Overdoses", 
                                "Specified Opioid Overdoses"))

ggplot(b2, aes(x = year, y = value, group(measure))) + 
  geom_point(aes(color = measure), size = 1.25) +
  geom_smooth(aes(color =measure), se = FALSE, size = 1.25) +
  scale_y_continuous(labels = scales::comma) +
  theme_tufte()  +
    xlab("Year") +
    ylab("Number of Deaths") +
  scale_colour_manual(values = c("gray5", "gray45", "gray80")) + 
  theme(legend.position=c(.4,.75),
        legend.title=element_blank())
ggsave(device = "png", 
       filename = "V:/seltzer/mortality/output/Predicted_Opioid_Deaths.png",
       width = 5,
       height = 5,
       dpi = 700)

#   scale_colour_scico_d(palette = "berlin")
# 
# +
#   scale_color_hue(l=40, c=35)
# 
#   
# scale_fill_gradientn(low = fake_scico[1], mid = fake_scico[2], high = fake_scico[3])
#   
# 
# 
# 
#   scale_color_gradient()
#   
#   
#   scale_fill_scico_d(palette = 'oslo')
#     
# scale_colour_viridis_d(option = "plasma")






mort_full3_pred <- bind_rows(pred_1999[[8]],
                             pred_2000[[8]],
                             pred_2001[[8]],
                             pred_2002[[8]],
                             pred_2003[[8]],
                             pred_2004[[8]],
                             pred_2005[[8]],
                             pred_2006[[8]],
                             pred_2007[[8]],
                             pred_2008[[8]],
                             pred_2009[[8]],
                             pred_2010[[8]],
                             pred_2011[[8]],
                             pred_2012[[8]],
                             pred_2013[[8]],
                             pred_2014[[8]],
                             pred_2015[[8]],
                             pred_2016[[8]],
                             pred_2017[[8]])
                             
  







###############################################################################
## 4.0 STATE ------------------------------------------------------------------
###############################################################################



mort_full4_sex_state_with_opioid_pred <- mort_full3_pred %>%
  group_by(state_fips, agecat, sex, year, ucod, pred_mix) %>%
  summarize(deaths = sum(death, na.rm = TRUE))


NCHS_sex_pop_state_drug <- NCHS_tot_pop_1 %>%
  rename(year = YEAR,
         state_fips = ST_FIPS2) %>%
  mutate(sex = ifelse(female==0, 1,
                      ifelse(female==1, 2, NA))) %>%
  group_by(state_fips, year, sex, agecat) %>%
  summarize(tot_pop = sum(pop, na.rm = TRUE)) %>%
  left_join(standardpop, by = "agecat")



## 4.2 STATE opioid drug death -------------------------------------------------

mort_full4_sex_state_opioid_pred <- mort_full4_sex_state_with_opioid_pred %>%
  filter(pred_mix == 1)

mort_full5_sex_state_opioid_pred <- mort_full4_sex_state_opioid_pred %>%
  filter(ucod == "X40" |
           ucod == "X41" |
           ucod == "X42" |
           ucod == "X43" |
           ucod == "X44" |
           ucod == "X60" |
           ucod == "X61" |
           ucod == "X62" |
           ucod == "X63" |
           ucod == "X64" |
           ucod == "X85" |
           ucod == "Y10" |
           ucod == "Y11" |
           ucod == "Y12" |
           ucod == "Y13" | 
           ucod == "Y14" ) %>%
  right_join(NCHS_sex_pop_state_drug) %>%
  mutate(age_crude_sex = (deaths/tot_pop)*100000,
         age_weight = age_crude_sex*weight) %>%
  arrange( state_fips, year, agecat, sex)

mort_full6_sex_state_opioid_pred <- mort_full5_sex_state_opioid_pred %>%
  group_by(state_fips, sex, year) %>%
  summarize(deaths_sex_opioid_state_pred = sum(deaths, na.rm=TRUE),
            tot_pop_sex = sum(tot_pop, na.rm = TRUE),
            cr_sex = sum(deaths_sex_opioid_state_pred/tot_pop_sex, na.rm = TRUE)*100000,
            asmr_sex_opioid_state_pred = sum(age_weight, na.rm = TRUE)) %>%
  filter(year > 1997) %>%
  select(-cr_sex, -tot_pop_sex)

# View(filter(mort_full6_sex_state_opioid_pred, year == 2008))

MORT_RATES_CLEAN <- "V:/seltzer/mortality/data/clean/mortality"


write.dta(mort_full6_sex_state_opioid_pred, paste0(MORT_RATES_CLEAN, "/sex_drugrates-RESTRICTED_keeponwinstat-state-opioid_predicted.dta"))
fwrite(mort_full6_sex_state_opioid_pred, paste0(MORT_RATES_CLEAN, "/sex_drugrates-RESTRICTED_keeponwinstat-state-opioid_predicted.csv"))


###############################################################################
## 4.0 COUNTY ------------------------------------------------------------------
###############################################################################



mort_full4_sex_county_with_opioid_pred <- mort_full3_pred %>%
  # mutate(county_fips = as.numeric(statecountyfips)) %>%
  group_by(statecountyfips, agecat, sex, year, ucod, pred_mix) %>%
  summarize(deaths = sum(death, na.rm = TRUE)) %>%
  mutate(county_fips = as.numeric(statecountyfips))


NCHS_sex_pop_county_drug <- NCHS_tot_pop_1 %>%
  rename(year = YEAR,
         county_fips = statecountyfips) %>%
  mutate(sex = ifelse(female==0, 1,
                      ifelse(female==1, 2, NA))) %>%
  group_by(county_fips, year, sex, agecat) %>%
  summarize(tot_pop = sum(pop, na.rm = TRUE)) %>%
  left_join(standardpop, by = "agecat")



## 4.2 county opioid drug death -------------------------------------------------

mort_full4_sex_county_opioid_pred <- mort_full4_sex_county_with_opioid_pred %>%
  filter(pred_mix == 1)

mort_full5_sex_county_opioid_pred <- mort_full4_sex_county_opioid_pred %>%
  filter(ucod == "X40" |
           ucod == "X41" |
           ucod == "X42" |
           ucod == "X43" |
           ucod == "X44" |
           ucod == "X60" |
           ucod == "X61" |
           ucod == "X62" |
           ucod == "X63" |
           ucod == "X64" |
           ucod == "X85" |
           ucod == "Y10" |
           ucod == "Y11" |
           ucod == "Y12" |
           ucod == "Y13" | 
           ucod == "Y14" ) %>%
  right_join(NCHS_sex_pop_county_drug) %>%
  mutate(age_crude_sex = (deaths/tot_pop)*100000,
         age_weight = age_crude_sex*weight) %>%
  arrange( county_fips, year, agecat, sex)

mort_full6_sex_county_opioid_pred <- mort_full5_sex_county_opioid_pred %>%
  group_by(county_fips, sex, year) %>%
  summarize(deaths_sex_opioid_county_pred = sum(deaths, na.rm=TRUE),
            tot_pop_sex = sum(tot_pop, na.rm = TRUE),
            cr_sex = sum(deaths_sex_opioid_county_pred/tot_pop_sex, na.rm = TRUE)*100000,
            asmr_sex_opioid_county_pred = sum(age_weight, na.rm = TRUE)) %>%
  filter(year > 1997) %>%
  select(-cr_sex, -tot_pop_sex)

# View(filter(mort_full6_sex_county_opioid_pred, year == 2008))

MORT_RATES_CLEAN <- "V:/seltzer/mortality/data/clean/mortality"


write.dta(mort_full6_sex_county_opioid_pred, paste0(MORT_RATES_CLEAN, "/sex_drugrates-RESTRICTED_keeponwinstat-county-opioid_predicted.dta"))
fwrite(mort_full6_sex_county_opioid_pred, paste0(MORT_RATES_CLEAN, "/sex_drugrates-RESTRICTED_keeponwinstat-county-opioid_predicted.csv"))
